fix any

+++ database/migrations/2018_01_26_032331_create_members_table.php
@@ -15,7 +15,7 @@ class CreateMembersTable extends Migration
     {
         Schema::create('members', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('user_id');
+            $table->tinyInteger('user_id');
             $table->string('name');
             $table->string('username')->unique();
             $table->string('password');
