+++ database/migrations/2018_01_26_061354_create_groups_table.php
@@ -15,7 +15,7 @@ class CreateGroupsTable extends Migration
     {
         Schema::create('groups', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('user_id');
+            $table->tinyInteger('user_id');
             $table->integer('name');
             $table->text('description')->nullable();
             $table->timestamps();
